services/gateway/requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv